Add tile-based algorithm for forest rendering
 - Reasoning: The current method of rendering LOD forests is to build a
giant quadtree of all trees on the entire map. For very large
forests, partitioning the whole map into nearby trees, faraway trees and
invisible trees affects performance. Additionally, the current method of
making a guess for the number of tree instances required on the GPU
results in either ugly gaps in the forest or wasted memory.

 - Core additions: The classes in tiny/draw/tiledhorde.h define the tiled
forest. Due to Renderable limitations, it is currently impossible to
share programs across Renderables and therefore I have not yet given
each Tile its own set of Renderable objects, but instead I copy the
positions of instances on each tile into a large array.

 - Supporting additions: I added tiny/algo/typecluster.h and
tiny/algo/gridmap.h to support the tile management. The former,
containing the TypeCluster and TypeClusterObject, aids memory management
and enables logarithmic complexity lookup (similar to a quadtree). The
latter, containing GridTile and GridMap, builds upon the TypeCluster to
provide a tile-based structure. Finally, I created a new test executable
to showcase and verify the new content. I also added a better heightmap,
512hills.png, since the tasmania heightmap's high altitude does not
result in trees near the starting location.

 - Future directions: I believe it would be more elegant to allocate an
array of instances on the GPU for each tile, rather than for the whole
world at once. This requires fundamental changes to the Renderable,
however, and I have chosen to postpone this for now.