Revert "Explained fork structure in README"
This reverts commit adebb8650fb3892e8beb7f38617ed07d1ece5454.

While the /master branch is still part of the pull request for the
TiledHorde I cannot yet push parts of it at will. So I reverted the
commit where I changed the README.

In the future I will make sure to use separate branches for pull
requests, and not the master branch, to avoid this issue.