Add functionality for freeing Renderables
That wasn't so hard.

But that's mostly a tribute to the nice engine. Anyways, here's the
changes and additions included in this commit:

I added a freeRenderable() function to the WorldRenderer. This in turn
calls freeRenderable of each Renderer, which frees all resources.

For doing this efficiently, I changed the Renderer's std::list of
BoundRenderables to an std::set. This allows logarithmic complexity
lookup (instead of linear complexity), which is important for deletion.
For traversal it shouldn't make a difference.

For the std::set to work, I added an operator< to the BoundRenderable
which works by comparing the Renderable* inside it. This enables another
wonderful construction, namely lookup of a BoundRenderable by creating a
dummy BoundRenderable based on the same Renderable*. This is used by
freeRenderable() to find the BoundRenderable to be deleted.

I also added shader program resource management by including a list of
Renderable*'s in each BoundProgram. Once no more Renderables exist using
the specific program, the program is deleted.