Move function bodies to .cpp and merge upstream
In this commit I cleaned up the header files for optimal compilation,
and before that I merged the upstream (which was identical to the
current 'font' except for the font-merge, which simply made the upstream
repository identical to 'font').

I also made a slight functional change: I added a function
ScreenIconHorde::eraseText() for erasing the text currently in the
ScreenIconHorde. This should aid removing text when so desired. I also
added calling this function to TextBox::setText() such that calling
setText() twice would work as expected (i.e. it would re-set the text).