Merge upstream changes for replacement pull request
I fought hard with Git but had to admit defeat in the end. I was unable
to explain to Git that I had taken my master branch (for which a pull
request was active), branched it into font, removed all the changes of
the pull request, and did a pull request for that. And then after that
was done I wanted my old master branch back.

Given that there was no way in which Git would let me get the
(previously committed and then deleted) additions for the TiledHorde
back into my local branch, I simply re-added them (since I had done some
backup-ing earlier) and hereby I am trying to force Git to digest it.

This commit should be properly up to date with the upstream (the
original tiny-game-engine repository) and be a suitable replacement to
the initial pull request that Git considered 'closed' earlier.