Allow new lines, suppress output, fix space wrap
A number of minor improvements are made here. Firstly, I allow making
new lines to allow paragraph or list structures to be rendered as a
single TextBox object. Secondly, I deleted two lines causing unnecessary
print statements.

Finally, I noticed that the word wrapping went wrong when the wrapping
was triggered by a space character, causing text to start away from the
left margin, as a consequence of using the yet-unused character beyond
the space at which wrapping occurs (which is not yet set if the space
itself is causing the wrapping).