Allow erasing of TextBox
A bit late but I just realised that one might want to clear the TextBox
contents (for the purpose of resetting what's in it). I thus added
TextBox::clear() which should do the necessary.

The clearing preserves the ScreenIconHorde that is used at its original
size, since one might want to re-add a text of similar length. If
freeing of the memory is really desired, one should simply delete the
TextBox object entirely.