Extend ScreenFillingSquare to allow all square sizes
The existing ScreenFillingSquare (see /tiny/draw/screensquare.h) has a
VertexBufferInterpreter that can only fill the entire screen, not a part
of it.

This commit adds a function setSquareDimensions() that allows the user
to reset the size of the screen square to some smaller rectangle that
only covers part of the screen.