Add functions for setting TextBox parameters
This adds two functions, setFontSize() and setAspectRatio(), that allow
setting the parameters 'size' and 'aspectRatio' of the TextBox. With
these included, the TextBox parameters are fully adjustable even after
construction of the object is complete.Add functions for setting TextBox
parameters

This adds two functions, setFontSize() and setAspectRatio(), that allow
setting the parameters 'size' and 'aspectRatio' of the TextBox. With
these included, the TextBox parameters are fully adjustable even after
construction of the object is complete.