Allow checking renderable index existence
Renderable indices (in the form of unsigned int's) must be unique for
freeing to be possible. Currently, the only way to 'test' uniqueness of
an index is to create inside a try-catch block and handle the exception
that is thrown on passing a renderable index that already exists.

In this commit, I add functions for checking the presence of a specific
index. This eliminates the risk of exceptions being thrown due to
ignorance regarding which indices are present in the Renderer. While it
can be claimed that the user is responsible for not passing the same
index twice, in complex programs it will undoubtedly be easier to simply
find a free index before trying to add a Renderable to the Renderer,
without having to maintain a duplicate of the Renderer::renderables()
map in the client application.